Including FOXSI-SMEX Ghost-rays
An example #7 is included showing the result of simulating 450k rays inpinging a FOXSI-SMEX Optics for a variety of Off-axis point sources.

The code tha generates this example of the Gallery does not include explicitely the SMEX-Prescription. Instead, a note in the header telles the user to email the PIs to get those parameters.